Move INetworkGraphAdapter out of inner class
